requirements.txt
Pillow
pydot
graphviz